订单接口修改

+++ src/main/java/gpps/service/ISubmitService.java
@@ -31,7 +31,7 @@ public interface ISubmitService {
 	 *             
 	 * @return 返回订单号
 	 */
-	public Integer buy(Integer productId, double amount) throws InsufficientBalanceException, ProductSoldOutException, InsufficientProductException, UnreachBuyLevelException;
+	public Integer buy(Integer productId, int amount) throws InsufficientBalanceException, ProductSoldOutException, InsufficientProductException, UnreachBuyLevelException;
 	public void confirmBuy(Integer submitId);
 
 	/**
